voice-to-action/python_backend/requirements.txt
# A2A Flow Dependencies
fastapi
uvicorn
pydantic
python-dotenv

# LangChain and AI
langchain
langchain-anthropic
langchain-mcp-adapters

# HTTP and async
aiohttp
requests